Remove registrertAnnenForelder from søkerinfo
